+++ sources/Device/src/Osci/Osci.cpp
@@ -25,7 +25,7 @@ static void UpdateFPGA();
 
 int Osci::addShift = 0;
 //  2нс 5нс 10нс 20нс 50нс
-const int Osci::Kr[TBase::Count] = { 50, 20, 10,  5,   2 };
+const uint Osci::Kr[TBase::Count] = { 50, 20, 10,  5,   2 };
 
 
 
@@ -120,9 +120,9 @@ static void UpdateFPGA()
 {
     bool needStop = false;
     
-    int number = (Osci::InModeRandomizer()) ? Osci::Kr[set.time.base] : 1;
+    uint number = (Osci::InModeRandomizer()) ? Osci::Kr[set.time.base] : 1;
 
-    for (int i = 0; i < number; i++)
+    for (uint i = 0; i < number; i++)
     {
         ContextOsci::ReadFlag();
     
@@ -304,7 +304,7 @@ Osci::StructReadRand Osci::GetInfoForReadRand(int Tsm, const uint8 *address)
             d = d;
         }
 
-        result.posFirst = index;
+        result.posFirst = static_cast<uint>(index);
     }
 
     return result;
